refactor(pd): inline ici_pd helpers into pathways_pd, drop ici backend
- Move _NPG_BATCH_MAX/_bucket_npg/slots_to_ordered_pages/migrate_reqs_p_to_d
  from ici_pd.py into pathways_pd.py (the only consumer being upstreamed).
- Drop the 'ici' choice for --pd-disaggregation and its scheduler branch.
- Rename [ici_pd] log prefix -> [pathways_pd].
